fix(backend): address PR #14 second review feedback
- Extract domain entity unit tests into PasswordResetTokenTest.kt (no Spring/DB needed)
- Add clarifying comment for @Transactional + @BeforeEach coexistence
- Replace !! with descriptive error() messages in E2E test helpers
